Adding multipart/byteranges to the never sniffed list.
Multipart range requests would allow bypassing CORB since the response
will have type multipart/byteranges instead of the resource's
original MIME type.

Blocking multipart responses is compatible with the web because the
renderer can only see into the result of a fetch for a multipart file
when the request is made with CORS. Media tags only make single-range
requests which will not have the multipart type.

Bug: 802836
Change-Id: Ib4ff928fdf9c604055f0d5ec52fcb5bee578099f