[LargestContentfulPaint] Add attributes and checks
This CL adds checks so that we do not expose rendering timestamp of
images which do not pass the Timing Allow Origin check. It also adds
attributes found in:
https://wicg.github.io/largest-contentful-paint/#sec-performance-largest-contentful-paint-candidate

The ImagePaintTimingDetector did not plumb the ImageResourceContent to
the ImageRecord, so this is fixed to allow LCP Calculator to make use of
it.

Bug: 965505
Change-Id: I019658dc6d130c8cea336f1ca70de6a32beebbb0